+++ app/Http/Controllers/API/StorefrontsController.php
@@ -6,6 +6,7 @@ use App\Models\Storefront;
 use App\Repositories\StorefrontsRepository;
 use App\Http\Requests\API\StorefrontsRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 
 class StorefrontsController extends APIBaseController
@@ -22,9 +23,13 @@ class StorefrontsController extends APIBaseController
     (
         Request $request,
         StorefrontsRepository $storefrontsRepository
-
     )
 	{
+	    $role = Auth::guard('api')->user()->can('storefronts_list');
+	    if (empty($role)) {
+	        return $this->sendError('无门店列表权限','403');
+        }
+
 	    $res = $storefrontsRepository->getStorefrontsList($request);
 	    return $this->sendResponse($res,'门店列表获取成功');
 	}
@@ -43,6 +48,11 @@ class StorefrontsController extends APIBaseController
     	StorefrontsRepository $storefrontsRepository
     )
     {
+        $role = Auth::guard('api')->user()->can('add_storefronts');
+        if (empty($role)) {
+            return $this->sendError('无添加门店权限','403');
+        }
+
     	$res = $storefrontsRepository->addStorefronts($request);
         return $this->sendResponse($res, '门店信息添加成功');
     }
@@ -75,6 +85,11 @@ class StorefrontsController extends APIBaseController
         Storefront $storefront
     )
     {
+        $role = Auth::guard('api')->user()->can('update_storefronts');
+        if (empty($role)) {
+            return $this->sendError('无更新门店权限','403');
+        }
+
         $res = $storefrontsRepository->updateStorefronts($storefront, $request);
         return $this->sendResponse($res,'更新成功');
     }
@@ -89,6 +104,11 @@ class StorefrontsController extends APIBaseController
      */
     public function destroy(Storefront $storefront)
     {
+        $role = Auth::guard('api')->user()->can('del_storefronts');
+        if (empty($role)) {
+            return $this->sendError('无删除门店权限','403');
+        }
+
         $res = $storefront->delete();
         return $this->sendResponse($res,'删除成功');
     }
